+++ integrations/linear/linear/utils.py
@@ -8,3 +8,7 @@ class ObjectKind(StrEnum):
     LABEL = "label"
     ISSUE = "issue"
     DOCUMENT = "document"
+    USER = "user"
+    PROJECT = "project"
+    TEAM_MEMBERS = "team-members"
+    CYCLE = "cycle"