+++ integrations/linear/webhook_processors/__init__.py
@@ -1,9 +1,15 @@
 from webhook_processors.issue_webhook_processor import IssueWebhookProcessor
 from webhook_processors.label_webhook_processor import LabelWebhookProcessor
 from webhook_processors.document_webhook_processor import DocumentWebhookProcessor
+from webhook_processors.user_webhook_processor import UserWebhookProcessor
+from webhook_processors.project_webhook_processor import ProjectWebhookProcessor
+from webhook_processors.cycle_webhook_processor import CycleWebhookProcessor
 
 __all__ = [
     "IssueWebhookProcessor",
     "LabelWebhookProcessor",
     "DocumentWebhookProcessor",
+    "UserWebhookProcessor",
+    "ProjectWebhookProcessor",
+    "CycleWebhookProcessor",
 ]